`ReparseMacroExpansions`: move code to top-level module again
This is done in a separate step from the previous commit, because that makes the
differ easier to review.
