+++ hs-bindgen/hs-bindgen.cabal
@@ -230,7 +230,6 @@ library internal
     HsBindgen.Frontend.Pass.PrepareReparse.Update
     HsBindgen.Frontend.Pass.ReparseMacroExpansions
     HsBindgen.Frontend.Pass.ReparseMacroExpansions.ForgetAnn
-    HsBindgen.Frontend.Pass.ReparseMacroExpansions.Intermediate.LanC
     HsBindgen.Frontend.Pass.ReparseMacroExpansions.IsPass
     HsBindgen.Frontend.Pass.ReparseMacroExpansions.IsPass.Msg
     HsBindgen.Frontend.Pass.ReparseMacroExpansions.LanC
